Refactoring
Be more consistent with existing preference pages
